refactor: move getting categories to custom hook
