+++ src/types/categories.ts
@@ -1,0 +1,5 @@
+export type FlatCategory = {
+  id: string;
+  categoryName: string;
+  nestingLevel: number;
+};
